[MNG-5760] Moved the -rf helper method, including the tests, back to MavenCli. The BuildResumer is now only focused on the --resume feature.
