+++ maven-core/src/main/java/org/apache/maven/execution/BuildResumer.java
@@ -21,8 +21,6 @@ package org.apache.maven.execution;
 
 import org.apache.maven.project.MavenProject;
 
-import java.util.List;
-
 /**
  * This class describes most of the logic needed for the --resume / -r feature. Its goal is to ensure newer
  * builds of the same project that have the -r command-line flag skip successfully built projects during earlier
@@ -56,21 +54,4 @@ public interface BuildResumer
      */
     void removeResumptionData( final MavenProject rootProject );
 
-    /**
-     * A helper method to determine the value to resume the build with {@code -rf} taking into account the edge case
-     *   where multiple modules in the reactor have the same artifactId.
-     * <p>
-     * {@code -rf :artifactId} will pick up the first module which matches, but when multiple modules in the reactor
-     *   have the same artifactId, effective failed module might be later in build reactor.
-     * This means that developer will either have to type groupId or wait for build execution of all modules which
-     *   were fine, but they are still before one which reported errors.
-     * <p>Then the returned value is {@code groupId:artifactId} when there is a name clash and
-     * {@code :artifactId} if there is no conflict.
-     *
-     * @param mavenProjects Maven projects which are part of build execution.
-     * @param failedProject Project which has failed.
-     * @return Value for -rf flag to resume build exactly from place where it failed ({@code :artifactId} in general
-     * and {@code groupId:artifactId} when there is a name clash).
-     */
-    String getResumeFromSelector( final List<MavenProject> mavenProjects, final MavenProject failedProject );
 }
