+++ maven-embedder/src/main/java/org/apache/maven/cli/MavenCli.java
@@ -1038,7 +1038,7 @@ public class MavenCli
             }
             else if ( project != null && !project.equals( sortedProjects.get( 0 ) ) )
             {
-                String resumeFromSelector = buildResumer.getResumeFromSelector( sortedProjects, project );
+                String resumeFromSelector = getResumeFromSelector( sortedProjects, project );
                 logBuildResumeHint( "mvn <args> -rf " + resumeFromSelector );
             }
 
@@ -1064,6 +1064,37 @@ public class MavenCli
         slf4jLogger.error( "" );
         slf4jLogger.error( "After correcting the problems, you can resume the build with the command" );
         slf4jLogger.error( buffer().a( "  " ).strong( resumeBuildHint ).toString() );
+    }
+
+    /**
+     * A helper method to determine the value to resume the build with {@code -rf} taking into account the edge case
+     *   where multiple modules in the reactor have the same artifactId.
+     * <p>
+     * {@code -rf :artifactId} will pick up the first module which matches, but when multiple modules in the reactor
+     *   have the same artifactId, effective failed module might be later in build reactor.
+     * This means that developer will either have to type groupId or wait for build execution of all modules which
+     *   were fine, but they are still before one which reported errors.
+     * <p>Then the returned value is {@code groupId:artifactId} when there is a name clash and
+     * {@code :artifactId} if there is no conflict.
+     * This method is made package-private for testing purposes.
+     *
+     * @param mavenProjects Maven projects which are part of build execution.
+     * @param failedProject Project which has failed.
+     * @return Value for -rf flag to resume build exactly from place where it failed ({@code :artifactId} in general
+     * and {@code groupId:artifactId} when there is a name clash).
+     */
+    String getResumeFromSelector( List<MavenProject> mavenProjects, MavenProject failedProject )
+    {
+        boolean hasOverlappingArtifactId = mavenProjects.stream()
+                .filter( project -> failedProject.getArtifactId().equals( project.getArtifactId() ) )
+                .count() > 1;
+
+        if ( hasOverlappingArtifactId )
+        {
+            return failedProject.getGroupId() + ":" + failedProject.getArtifactId();
+        }
+
+        return ":" + failedProject.getArtifactId();
     }
 
     private void logSummary( ExceptionSummary summary, Map<String, String> references, String indent,
